Create schedule_job_probe.py
Uhm, I have no clue if I am doin' it rite yo. This is why we should not rewrite our code in the wee hours of the morning in a panicked rush to commit. Anyway, even if I am doing it disgustingly wrong, you get the idea. We use the probe.cfm XSS to upload a "stager" fileuploader, then use that to upload fuse or w/e.